Refactor project layout and add docs
