+++ README.md
@@ -44,11 +44,12 @@ Just this:
 1. Install **Python 3**.
 2. Run the interpreter demo on the boot example:
 
-```bash
-python interpreter_example.py rules/examples/boot.trx
-```
+ ```bash
+ python -m trxcore.interpreter_example rules/examples/boot.trx
+ ```
 
 3. Modify an example file such as `hello_nova.trx` and run the script again to see the structure change.
+4. For more commands and API details, see `docs/user_guide.md` and `docs/api_reference.md`.
 
 ## ✦ Examples
 
@@ -79,7 +80,7 @@ It tries to be **a mirror for structure**.
 
 ## ✦ Benchmarks & Further Reading
 
-Run `python benchmark_extended.py` to time the parser and translator across all
+Run `python -m trxcore.benchmark_extended` to time the parser and translator across all
 example programs. A summary of the results appears in
 `docs/impact_report.md`.
