+++ requirements.txt
@@ -1,0 +1,1 @@
+# All functionality uses the Python standard library
